% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ancova_data}
\alias{ancova_data}
\title{ancova_data is a test dataset for ANCOVA.}
\format{
A data frame with 76 rows and 3 variables:
\describe{
  \item{group}{group information}
  \item{W1D1}{covariable}
  \item{diffW24}{outcome}
  ...
}
}
\source{
\url{}
}
\usage{
ancova_data
}
\description{
A dataset containing three variables for ANCOVA
}
\keyword{datasets}
