% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RROR.R
\name{calcOddsRatio}
\alias{calcOddsRatio}
\title{Title}
\usage{
calcOddsRatio(mymatrix, alpha = 0.05, referencerow = 2, quiet = FALSE)
}
\arguments{
\item{mymatrix}{a two dimension matrix}

\item{alpha}{P value}

\item{referencerow}{disease row}

\item{quiet}{quite or not}
}
\value{
string
}
\description{
Title
}
\examples{
mymatrix <- matrix(c(156, 9421, 1531, 14797), nrow = 2, byrow = TRUE)
calcOddsRatio(mymatrix, alpha = 0.05, referencerow = 2 )
}
