% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSampleSizeForRCT.R
\name{calcPowerForRCT}
\alias{calcPowerForRCT}
\title{Title}
\usage{
calcPowerForRCT(alpha, piT, piC, n)
}
\arguments{
\item{alpha}{the significance level}

\item{piT}{the estimated incidences of the disease in the disease}

\item{piC}{the estimated incidences of the disease in the control}

\item{n}{number for each group}
}
\value{
string
}
\description{
Title
}
\examples{
calcPowerForRCT(alpha=0.05, piT=0.2, piC=0.3, n=250)
}
