% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RROR.R
\name{calcRelativeRisk}
\alias{calcRelativeRisk}
\title{Calculate Relative Risk}
\usage{
calcRelativeRisk(mymatrix, alpha = 0.05, referencerow = 2)
}
\arguments{
\item{mymatrix}{a two dimension matrix}

\item{alpha}{P value}

\item{referencerow}{disease}
}
\value{
string with value and ci
}
\description{
Calculate RR value in table
}
\examples{
mymatrix <- matrix(c(156, 9421, 1531, 14797), nrow = 2, byrow = TRUE)
calcRelativeRisk(mymatrix)
}
