% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSampleSizeForRCT.R
\name{calcSampleSizeForRCT}
\alias{calcSampleSizeForRCT}
\title{calcSampleSizeForRCT}
\usage{
calcSampleSizeForRCT(alpha, gamma, piT, piC, p = 0)
}
\arguments{
\item{alpha}{the significance level}

\item{gamma}{the power that you want}

\item{piT}{the estimated incidences of the disease in the disease}

\item{piC}{the estimated incidences of the disease in the control}

\item{p}{lost follow-up ratio}
}
\value{
string
}
\description{
calcSampleSizeForRCT
}
\examples{
calcSampleSizeForRCT(alpha=0.05, gamma=0.80, piT=0.15, piC=0.2)
#tatol sample should be doubled
}
