% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tarones.R
\name{calcTaronesTest}
\alias{calcTaronesTest}
\alias{Tarones}
\title{Tarones test for Stratified OR}
\usage{
calcTaronesTest(mylist, referencerow = 2)
}
\arguments{
\item{mylist}{A list of matrix Stratified by a Variable}

\item{referencerow}{Unexposed row}
}
\value{
string with results
}
\description{
\code{calcTaronesTest} is ued to test whether the OR in different strata are different
}
\examples{
mymatrix1 <- matrix(c(4,5,5,103),nrow=2,byrow=TRUE)
colnames(mymatrix1) <- c("Disease","Control")
rownames(mymatrix1) <- c("Exposure","Unexposed")
mymatrix2 <- matrix(c(10,3,5,43),nrow=2,byrow=TRUE)
colnames(mymatrix2) <- c("Disease","Control")
rownames(mymatrix2) <- c("Exposure","Unexposed")
mylist <- list(mymatrix1,mymatrix2)
calcTaronesTest(mylist)
\dontrun{
calcTaronesTest(mymatrix1)
}
}
