% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{ci_diff_mean}
\alias{ci_diff_mean}
\title{Confidence interval for a difference in means}
\usage{
ci_diff_mean(
  x,
  y,
  var.equal = FALSE,
  conf.level = 0.95,
  methods = c("t.test", "pairwiseCI")
)
}
\arguments{
\item{x}{numbers in sample1}

\item{y}{numbers in sample2}

\item{var.equal}{TRUE or FALSE}

\item{conf.level}{confidence level of the returned confidence interval}

\item{methods}{'t.test', 'pairwiseCI'}
}
\value{
a matrix
}
\description{
Confidence interval for a difference in means
}
\examples{
ci_diff_mean(x = c(1,2,3), y = c(3,4,5), methods = 't.test')
}
