% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{ci_diff_median}
\alias{ci_diff_median}
\title{Confidence interval for a difference in median}
\usage{
ci_diff_median(formula, data, alternative = "two.sided", conf.level = 0.95)
}
\arguments{
\item{formula}{A formula of the structure response ~ treatment for numerical variables}

\item{data}{A data.frame containing the numerical response variable and the treatment and by variable as factors}

\item{alternative}{Character string, either "two.sided", "less" or "greater"}

\item{conf.level}{The comparisonwise confidence level of the intervals}
}
\value{
a matrix
}
\description{
Confidence interval for a difference in median
}
\examples{
df <- data.frame(group = rep(c('a', 'b'),5), value = sample(1:10, 10))
ci_diff_median(value ~ group, data = df)
}
