% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{ci_diff_proportion}
\alias{ci_diff_proportion}
\title{Title}
\usage{
ci_diff_proportion(
  x1,
  n1,
  x2,
  n2,
  conf.level = 0.95,
  sides = c("two.sided", "left", "right"),
  method = c("ac", "wald", "waldcc", "score", "scorecc", "mn", "mee", "blj", "ha",
    "hal", "jp")
)
}
\arguments{
\item{x1}{number of success for the first group}

\item{n1}{number of trails for the first group}

\item{x2}{number of successes for the second group}

\item{n2}{number of trails for the second group}

\item{conf.level}{confidence level, defaults to 0.95.}

\item{sides}{a character string specifying the side of the confidence interval,
must be one of "two.sided" (default), "left" or "right".}

\item{method}{one of "wald", "waldcc", "ac", "score", "scorecc", "mn", "mee", "blj", "ha", "hal", "jp".}
}
\value{
A matrix with 3 columns containing the estimate, the lower and the upper confidence intervall.
}
\description{
Title
}
\examples{
ci_diff_proportion(x1=56, n1=70, x2=48, n2=80)
}
