% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{ci_mean}
\alias{ci_mean}
\title{calculate mean CI for single vector}
\usage{
ci_mean(x, conf.level = 0.95)
}
\arguments{
\item{x}{numeric vector}

\item{conf.level}{confidence level, defaults to 0.95.}
}
\value{
mean CI tibble, same as t.test
}
\description{
calculate mean CI for single vector
}
\examples{
ci_mean(mtcars$mpg)
}
