% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{ci_proportion}
\alias{ci_proportion}
\title{calculate ci for proportion}
\usage{
ci_proportion(x, n = NULL, conf.level = 0.95)
}
\arguments{
\item{x}{vector of positive integers representing the number of occurrences of each class, or number of
successes(n is needed)}

\item{n}{number of trials.}

\item{conf.level}{confidence level, defaults to 0.95.}
}
\value{
A vector with 3 elements for estimate, lower confidence interval and upper for the upper one.
}
\description{
calculate ci for proportion
}
\examples{
ci_proportion(c(2,3,4))
ci_proportion(x = 37, n = 43)
}
