% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI.R
\name{ci_proportion_ratio}
\alias{ci_proportion_ratio}
\title{Confidence interval for a ratio in proportion}
\usage{
ci_proportion_ratio(
  formula,
  data,
  alternative = "two.sided",
  conf.level = 0.95
)
}
\arguments{
\item{formula}{A formula of the structure response ~ treatment for numerical variables}

\item{data}{A data.frame containing the numerical response variable and the treatment and by variable as factors}

\item{alternative}{Character string, either "two.sided", "less" or "greater"}

\item{conf.level}{The comparisonwise confidence level of the intervals}
}
\value{

}
\description{
Confidence interval for a ratio in proportion
}
\examples{
df <- data.frame(group = rep(c('a', 'b'),5), success = sample(1:10, 10), failure = sample(1:10, 10))
ci_proportion_ratio(cbind(success, failure) ~ group,data = df)
}
