% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinstats-package.R
\docType{package}
\name{clinstats-package}
\alias{clinstats}
\alias{clinstats-package}
\title{clinstats: Clinical Research Statistics Toolkit}
\description{
A compact set of statistical utilities that recur in clinical and
translational research: ROC summaries, odds ratios, univariable and
multivariable Cox and logistic regression tables, survival cut points,
resampling-based Cox screening, TCGA clinical extraction, and
expression-matrix cleaning. Curated colorectal cancer clinical datasets
assembled from public repositories are included, together with a Quarto
analysis report template.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gflab/clinstats}
  \item Report bugs at \url{https://github.com/gflab/clinstats/issues}
}

}
\author{
\strong{Maintainer}: Feng Gao \email{gaofeng21cn@gmail.com}

Authors:
\itemize{
  \item Feng Gao \email{gaofeng21cn@gmail.com}
}

Other contributors:
\itemize{
  \item Lina Zhu [contributor]
  \item Wei Wang [contributor]
}

}
\keyword{internal}
