% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{df is a dataset for test.}
\format{
A data frame with 96 rows and 4 variables:
\describe{
  \item{density}{different density}
  \item{block}{different block}
  \item{fertilizer}{fertilizer}
  \item{yield}{outcome}
  ...
}
}
\source{
\url{}
}
\usage{
df
}
\description{
A dataset containing three category variable and one continue variable
}
\keyword{datasets}
