% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseOddsDiseaseRegression.R
\name{doseOddsDiseaseRegression}
\alias{doseOddsDiseaseRegression}
\title{Title calculate OR value by lR}
\usage{
doseOddsDiseaseRegression(mymatrix, referencerow = 1)
}
\arguments{
\item{mymatrix}{a two dimension matrix}

\item{referencerow}{which disease row as reference}
}
\value{
string with results
}
\description{
Title calculate OR value by lR
}
\examples{
mymatrix <- matrix(c(35,82,250,293,196,190,136,71,32,13),nrow=5,byrow=TRUE)
colnames(mymatrix) <- c("Disease","Control")
rownames(mymatrix) <- c("2","9.5","19.5","37","50")
doseOddsDiseaseRegression(mymatrix)
}
