% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseSpecificOddsRatios.R
\name{doseSpecificOddsRatios}
\alias{doseSpecificOddsRatios}
\title{doseSpecificOddsRatios}
\usage{
doseSpecificOddsRatios(mymatrix, referencerow = 1)
}
\arguments{
\item{mymatrix}{a two column matrix with level variable}

\item{referencerow}{reference row}
}
\value{
string
}
\description{
doseSpecificOddsRatios
}
\examples{
mymatrix <- matrix(c(35,82,250,293,196,190,136,71,32,13),nrow=5,byrow=TRUE)
colnames(mymatrix) <- c("Disease","Control")
rownames(mymatrix) <- c("2","9.5","19.5","37","50")
doseSpecificOddsRatios(mymatrix)
}
