% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSampleSizeForRCT.R
\name{calcSampleSizeForRCT}
\alias{calcSampleSizeForRCT}
\title{Title}
\usage{
calcSampleSizeForRCT(alpha, gamma, piT, piC, p = 0)
}
\arguments{
\item{alpha}{the significance level}

\item{gamma}{the power that you want}

\item{piT}{the estimated incidences of the disease in the disease}

\item{piC}{the estimated incidences of the disease in the control}

\item{p}{if we estimate that 10% of the people are likely to be lost to follow-up, then p=0.1}
}
\value{
a string
}
\description{
Title
}
\examples{
calcSampleSizeForRCT(alpha=0.05, gamma=0.80, piT=0.15, piC=0.2)
#This tells us that the sample size required in each group is 1214 people,
#so overall we need 1214*2=2428 people in the randomised control trial.
}
